added filter

+++ database/migrations/2021_12_24_074520_create_reviews_table.php
@@ -15,7 +15,7 @@ class CreateReviewsTable extends Migration
     {
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
-            $table->integer('rating');
+            $table->decimal('rating', 2, 1);
             $table->text('comment');
             $table->bigInteger('author_id')->unsigned();
             $table->bigInteger('book_id')->unsigned();
